Rename some functions and improve docs
Rename some of the confusing function names and add further
clarification in the docs.

Changed some of the bitstring() to binary() since that is what the
values really are after all. The decode/1 still uses bitstring because
changing to binary fails to match the input, for whatever reason.

rebar3 ct passes